BAU: Bump gradle image version to 7.3.3 and remove SHA's
The versions of the gradle and openjdk images being referred to with
their SHAs were for amd64 architecture. These images didn't work on an
M1 which is arm64.

As we now have a mix of architectures being used for development we
should remove these SHAs and let Docker decided which version of the
image to pull.

This also stops using a gradle image and instead uses the openjdk image
with gradle wrapper. This should be easier to maintain going forwards.