PYI-698: Use ClientAuthenticationVerifier for auth JWT verification
Use the client authentication verifier supplied in the Nimbus library.

This will save having to use a bunch of custom validation code, and
means it will be easier to have a standardised approach across system
components.

We need to implement the `ClientCredentialsSelector` class, so that we
can tell the verifier how to fetch the public key for each client. The
implementation added here just pulls the certificate from the config
class we're using. Other implementations, such as in core-back, could
pull them directly from the param store.