PYI-677 Relax Redirect Url Validation on Stubs
To enable the simpler reusing of stubs across Developer environments
accept redirect urls that include the London PaaS domain. The validation
could be tighter since this will accept a URI where the PaaS domain
simply appears in the path but these are stubs and including the extra
logic to guard against this seems unnecessary at present.