PYI-731: Return Verified Credential from CRI stub
Returns a Verified Credential from CRI stub in JWT format. See here for
details: https://www.w3.org/TR/vc-data-model/#json-web-token

As this is this stub we're just returning all of the shared attributes
received. These should be in the format defined in our schema. It can be
found here: https://vocab.london.cloudapps.digital/

We're also including any data sent in via the form on the stubs
frontend.

The JWK is signed using an EC key.

This also DRYs up some of the tests.