PYI-731: Update credential issuing route
To be in line with vc-api spec, the route for issuing credentials should
be a POST to /credentials/issue. This also allows us to pass the
unsigned JWT with the sub ID in it as part of the body of the request
from core.

It also stipulates that we should return a 201.

API spec here: https://w3c-ccg.github.io/vc-api/#issuing