PYIC-905: Make CRI algo configurable
This adds some logic to allow the signatures for the CRIs to be
configurable between RS256 and ES256.

It also makes the HTTP method used when calling the issue credentials
endpoint configurable between GET and POST.

These changes are to allow the core-stub to run against the passport
CRI whist not breaking the interface with the KBV teams CRIs.