PYIC-1009 Permit API Gateway Domains for Callback URL
We share the CRI stubs across developer environments. The validation on
the callback url was previously relaxed to allow all PaaS domains in
order to permit this sharing. Now core-front is being migrated to AWS
and is invoked behind an API Gateway we need to permit AWS API Gateway
like domains. When the stubs are deployed we will not know all of the
API Gateway ids for each of the developer environments which can change
as their environments are created and destroyed.

Once we have DNS set-up (and if we apply it to development
environments) this validation can be tightened, or post MVP a more
sophisticated check added.