BAU: Include `aud` claim in VC JWT
The core is now validating the claims included in the JWT returned by
the CRI. This includes validating the audience claim, which we weren't
including.

This updates the CRI stub to store the client ID along with the
credential. This is so it can be retrieved easily in the resource
endpoint and used to fetch the correct audience value from the client
configs. We then use that value in the JWT claims.