PYIC-1125: Return requested error from token endpoint
This allows the requested error to be persisted and returned from the
token endpoint. We needed a way to remember state a link it to a
specific journey for when the user comes back to the token endpoint, and
the auth code seems like a good way to do that.

This also updates the template to use a combination of radio buttons and
a drop down - hopefully it's easier to use.

The tests are fairly minimal and don't cover all use cases. The new
service is also being tested implicitly rather than directly. It's also
not being tested across an actual journey between the auth and token
endpoints. It would require a functional test with an app which is more
effort that it's probably worth right now.

One things this piece of work doesn't do is allow to the return of an
error response on the initial call to the auth endpoint. It's possible
for an error to happen, but we'd need a way to "prime" the CRI before
getting to it's auth page. Some sort of backchannel request. Maybe this
could be looked at in future.