PYIC-2024: Generate random IDs for app journey users
Once we start to use the stored CIs in the core, using the same user
ID for functional tests that generate CIs will break.

The core has been updated to send any user whos ID starts with a certain
prefix on an app journey. This way we can use a new user ID and still
trigger an app journey.