PYIC-2433 Refactor of Paths
We could be deploying multiple stubs, so with this in mind the paths
have been refactored to allow for this to happen.

The pipelines will need updating as well though.

moved passport private api key to secret env variable, added permissions for
reading from secrets manager.

updated gha to use the new paths and made it do a granular lookup to
find out what has changed specifically for this stub deployment.

	renamed:    .github/workflows/ipv-core-stub.yml -> .github/workflows/passport-core-stub.yml
	renamed:    di-ipv-core-stub/deploy/template.yaml -> di-ipv-core-stub/deploy/passport/template.yaml