PYIC-2802: Return jti claim in VCs set to UUID
As per [ADR 77](https://github.com/alphagov/digital-identity-architecture/blob/main/adr/0079-unique-identifier-for-verifiable-credential.md) this adds a jti claim to VCs returned by the CRI stub. The value is a UUID.

This is a second attempt at doing this. Initially I was just setting the
value with `UUID.randomUUID()`. The problem with this is the value was a
UUID object rather than a string. When serialized to go over the wire it
was getting converted to an empty JSON object, rather than a string.
This was correctly rejected by core-back.

This correctly sets the value in the VC specifically as a string,
allowing it to be serialized correctly.

This was initially missed as the test exercising the generation of the
VC wasn't serializing/deserializing the VC. I've added this step, which
has required some adjusting of the expected values. Parsing of a JWT
will set the audience value as a list, even if there is only one value.
Also, the expiration and not before times are in milliseconds, rather
than seconds.