PYIC-3386: Add cross account perms to cimit stubs
When the cimit stub is deployed to the stubs accounts (build and prod),
they will need to be given a lambda invoke permission, with the core
accounts as the principal. Otherwise core won’t be able to call them.

This hasn’t been a problem before as we’ve been testing entirely in the
dev account, so no cross account access.

We should give the stubs build and prod deploy the perms to allow dev
and build core envs to invoke them. This will allow a more flexible
approach (we can hook up to whatever stub we want from either core
account), whilst keeping the stub reasonably locked down. Another
approach is to give our AWS org permissions to invoke the lambdas, but
that leaves them open to anyone at GDS.