BAU: Fix css assets path
The logs for orch stub are very noisy. There are a lot of errors coming
from Spark due to requests for assets that haven't been mapped. This is
due to the CSS files trying to load resources from the wrong place.
They're looking in `/assets/...` when they should be looking in
`/gds/assets...` instead.

I'm sure that updating these CSS files by hand is the wrong thing to do.
I don't know what the correct way of fixing this is. But this works.