BAU: URL encode userId when using CiMit management API
We've seen some failures when mitigating CIs from the CRI stub. This
happens when the user ID is in the format `urn:uuid:kjddf....`.
This is the correct format and is the format used by the orch stub if
you don't override it.

However the user ID is used as part of the URL when hitting the CiMit
stub management endpoint and having the colons breaks things.

This URL encodes them so the user ID is safe.