PYIC-3781: Update StubManagementHandler for mitigations
This updates the stub management handler to add a new pending mitigation
item to a new table when it receives a mitigation request. This is
rather than updating the CI item itself.

This is to allow us to only mark a CI as mitigated once we actually post
a VC to the mitigations endpoint on the stub. This will replicate the
actual system more closely.