PYIC-3781: Check pending mitigations in post-mitigations lambda
This adds logic to the post-mitigation lambda to check the new pending
mitigations table, and then mitigate a users CI if there is both a
pending mitigation and a CI item to mitigate.

This also checks if the original mitigation request was a POST or PUT
and then either adds to or replaces the CIs mitigations accordingly.