PYIC-3782: Send JTI in mitigation request from CRI stub
The cimit stub is now expecting to receive a JTI (jwt ID) in the
requests to mitigate CIs. This is so it can do the actual mitigation at
the point the VC is received on the post-mitigations endpoint.

This reorders the logic in the CRI stub to generate the signed JWT (and
so the JTI) for the VC before storing it. This means it doesn't need to
get generated by the credential handlers. And it means we have the JTI
available to send in the request to the cimit stub.