BAU: Include esbuild in ticf stub prod dependencies
esbuild is required to build the project. If you don't have it installed
globally already running `sam build` will fail with the following:

```
Build Failed Error: NodejsNpmEsbuildBuilder:EsbuildBundle - Esbuild
Failed: Cannot find esbuild. esbuild must be installed on the host
machine to use this feature. It is recommended to be installed on the
PATH, but can also be included as a project dependency.
```

This adds the dependency to the production dependencies. This is because
as of version 1.56.0 of the SAM cli `sam build` will install
dependencies with the `production` flag. esbuild will bundle the code so
won't actually be included in the final package. See here for details
https://github.com/aws/aws-sam-cli/issues/4183#issuecomment-1234862956