BAU: Fetch client config from SSM
This creates a new config service for fetching client config from the
parameter store, rather than loading it in from a file.

This moves us towards how we manage config in the other components, and
should make the set up for the CRI stub easier to understand.

The new ConfigService is currently only used for client config, but
could be extended for all config in future work.

It's been implemented as a static class. This means it can just be
called where it's needed and avoids having to inject it everywhere. This
does make it harder to test, and a stub client has been created to allow
some testing to happen.