BAU: Fix orch-stub vtr parsing
The value entered in the input on the front page of orch stub was being
read as a single array with one string in. This chops the value up on a
comma delimiter to allow multiple values.

We were also wrapping the array inside a list when setting it on the
claimsSet. This just sets the array as the value, which is what we
acutally want.

This also removes the conditional check on if the value exists or not.
If the input box is left empty then there is still a value of empty
string, so it was a pointless check.

It also removes the default of P2 that was being set in the backend.
This will allow to test the case where no vtr is passed in.