PYIC-5081: Update CRI stub subject data
This updates the canned subject data for the passport and driving
licence CRIs.

For passport it adds the `icaoIssuerCode` property to the document
information. This will be needed by the cimit stub when returning
document types.

This also adds a missing issued number to the driving license data. This
isn't really needed by anything for now.