PYIC-5082: Store CI items individually
This updates the CimitStubTable to use a different primary key. This is
to allow us to store multiple CIs with the same code.

This is neccesary due to the way the real cimit service will be
returning CIs to us.

The new sortKey is a composite of `<ciCode>#<issuanceDate>`. This will
allow unlimited items for a given sort code (as long as they're not
issued in the same second), and still allow us to query for them with
just the CI code.

This change involves removing the name we set for our tables. Setting a
custom name makes CloudFormation very grumpy if the resources needs
replacing. Better to just let CloudFormation name the thing.