BAU: Fix txn calues in cimit VC
The cimit VC contains a list of txn IDs at the evidence level, and at
the individual CI item level.

The evidence level is a set of all the txn's in the CI items.

This has implmented the CI items as a list of the txns from all the CIs
that have formed that item - so from all deduplicated ones. Waiting on
confirmation from SPoT if that's actually correct.