PYIC-5401: Add users to passport and DL CRI
This adds Kenneth as a driving licence CRI user, and Alice as a passport
user.

This will allow the new mitigation routes to be run with a consistent
user. It is possible to run the journeys without this, as we don't do
correlation with failing VCs, it will make for a less confusing
experience.