PYIC-6182: Use domain object for auth response
This updates the AuthHandler#generateResponse method to parse a domain
object from the request to use, rather than accessing the query params
directly. This is to allow us to use a similar domain object when
handling an API request to the headless endpoint. The params receieved
in the usual flow and the API endpoint will be slightly different, and
using an object should mean we can reuse the same logic for generating
response on both modes.

This also removes all the logic around expiry dates for VCs, as we don't
use them.