PYIC-5874: Add new api endpoint
This adds a new endpoint for API based auth requests. An AuthRequest is
deserialized from the incoming JSON body, and then fed into the same
logic that the form based requests use.

The logic has been extracted to a method so it can be shared.

There's been some renaming of the fields in the AuthRequest object, and
some null checks added, but it's mostly exactly the same logic.
	modified:   di-ipv-credential-issuer-stub/src/main/java/uk/gov/di/ipv/stub/cred/handlers/AuthorizeHandler.java