PYIC-5874: Set content type and add tests
This updates the api auth endpoint to set the content type of the
response to application/json

It also renames cis to ci to be more inline with the evidence json.

Also now loads the F2F queue URL dynamically, rather than from a static
constant. It's easier to test.

Also now uses an http client to send message to the F2F queue rather
than the nimbus lib. Easier to test and more consistant.

Adds a bunch of tests for the api auth endpoint, and refactors the
structure of the tests to hopefully make them more obvious.