BAU: Don't include shared attributes for user asserted CRIs
This prevents the shared attributes from being included in the response
from the CRI stub if the CRI type is user asserted.

The only user asserted CRIs are address and claimed identity - neither
of which use the shared attributes in the real world.

This is being done to fix an issue with the COI journey. If a user is
changing their name only and not their address we will reuse their
address VC.

Before this change, the address VC from the stub contained the users
name. On the COI journey, the users journey would fail after updating
their name as the names in all the VC no longer correlated.