PYIC-6218: Include evcs access token in orch-stub JAR
This adds logic to generate an access token for the evcs storage system
and include it as a claim in the JAR. The evcs system requires this
before reading anythign from it. Core will parse it and send it in
requests to evcs.

This change introduceds two ways of generating a token.

If the stub is being used against the staging environment it will call
an API provided to us by the accounts bravo team. That will return to us
a token signed by them that will be accepted by the staging deploy of
the evcs system that we will be integrating with. We provide the user ID
and a ttl to the API.

If the stub is running in any other env (dev and build) it will generate
a sign a token itself. It's using the private key that matches the
public key that our stub of evcs is using to validate access tokens. Our
stub only validates the signature and that the `sub` attribute of the
JWT matches the user ID sent in the get request to it. So we don't need
to include anything else in the JWT.