BAU: Refactor cri-stub resource ID
We assign a random ID to link auth codes, access tokens, and the
resource associated with them.

We were setting this as a value in the form on the page and then
passing it back in. This meant that we also needed to provide a value in
API calls to the CRI stub.

We can simplify all this and just generate a UUID at the point we need
it.