PYIC-7081: Remove powertools from cimit stub
The powertools-tracing lib had a dependency on an old version of
ion-java which had a vulnerability (see link below). To make things more
complicated AWS have changed the package name, so forcing a newer
version that's patched (which does exist), is not possible.

Turns out we're not actually using either powertools tracing or logging
in the stub, so the easiest thing is to just remove the offending
dependencies.

https://github.com/govuk-one-login/ipv-stubs/security/dependabot/33