PYIC-7084: Bump oauth2-oidc-sdk to 11.13
The old version of the oauth-oidc-sdk we were using was pulling in a
vulnerable version of json-smart - see link.

This bumps it to the latest and greatest which doesn't suffer.

Had to make a few changes to the CRI stub, as the JSON parser that came
with the lib is no longer present, so using Jackson instead.