PYIC-7097: Upgrade awssdk and stub-oauth-client
The older version of the awssdk were pulling in an old version of
fast-xml-parser which had a vulnerability.

The stub-oauth-client lib dependencies were also using the old version.
The lib has already been updated upstream, so just running npm install
has brought in the newer version. This changes the git address to the
new GitHub org from alphagov - the repo has been migrated.

https://github.com/govuk-one-login/ipv-stubs/security/dependabot/7