PYIC-7198: Add response delay option
We want to be able to test slow resonses from ticf in core. This adds a
configuration option to the management api that causes ticf stub to
sleep for a configured period.

It also adds prettier which seemed to be missing and runs the `fix`
script. Shoould have done that in a separate commit. Sorry.