BAU: Clean up CIMIT stub template permissions
We now only call the CIMIT stub via API gateway. So we can remove all
the invoke permissions for the core accounts.

I did think that we shouldn't need to define the invoke permissions for
the internal API gatway on the lambdas, but removing it also removes API
gateways permissions. This is confusing, as core-back internal API, for
example, doesn't define permissions like this and it works fine. There
seems to be some issue with importing an OpenAPI spec and those perms
being create (found people reporting similar issues online). I'm just
leaving them in for now.