PYIC-4589: Remove env from orch stub parameter locations
The orch-stub CloudFormation template pulls values for env vars from the
parameter store. The path to the values includes an environment slug.
This slug is always set to "build" which is confusing.

When defining a parameter in CloudFormation using the
`AWS::SSM::Parameter::Value` type the default value has to be a literal
string. You can't use the !Sub function to parameterise it for different
envs for example.

The values for the parameter store parameters are defined in the
stubs-common-infra repo and are defined per env (build and prod). We
also define them in the dev config files in core-common-infra so we can
deploy the orch stub to the dev account.  So we do already have some
separation of config per env.

Setting config values as env vars in the ECS cluster means we don't need
to worry about caching, and retrieving the values at runtime should be
practically instantaneous. This is important for when we run the perf
tests, as the orch-stub can be a bottle neck. I don't really want to
change this and add another variable when analysing the perf tests.

As such, this PR just removes the use of the env slug. This should help
to remove confusion about where the values are coming from.