BAU: Strip free text inputs to orch stub
The most of the freetext inputs to orch-stub weren't being stripped.
This is most notably a problem if you copy a userId from Splunk and
paste it in - it comes with a whitespace.

The result is that the user ID in the EVCS signed JWT access token
doesn't match the user ID that gets sent by core to EVCS (which probably
does strip it somewhere). And you get an access denied error from EVCS.