PYIC-7581: Add `/migration/` endpoint to evcs stub
As part of the plans to reconcile the VCs migrated into EVCS, the EVCS
system will be opening a temporary endpoint for us to fetch VCs without
an access token. This adds the same to our stub to allow us to test the
process in dev and build.

EVCS actual implementation can be found here: https://github.com/govuk-one-login/di-id-reuse-storage/pull/1047