BAU: Noop to run the pipeline
The most recent deploy of orch stub has resulted in a strange half
state. The deploys were successful, but the ECS services tasks are stuck
on an old, inactive version. So the new changes haven't been applied. If
you try to manually update it ECS gives you an error about the
deployment not being found. This happened in both stubs-build and
stubs-prod.

I've deleted the orch-stub stack in stubs-build and rerun the pipeline,
but now the stack won't deploye. The ECS service fails as the
TaskDefinition can not be blank. Even though it creates one.

This is an attempt to generate another entirely new task definition, to
see if it can unstick it.